requirements.txt
flask
google-cloud-bigquery
pandas
notion-client
python-dotenv
db-dtypes
